requirements.txt
PiPlate
numpy